requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.78.0
frozenlist==1.3.3
h11==0.14.0
idna==2.8
multidict==6.0.4
openai==0.26.4
pydantic==1.9.2
PyJWT==2.4.0
pymongo==3.12.3
python-dotenv==0.21.1
PyYAML==5.3.1
requests==2.22.0
sniffio==1.3.0
starlette==0.19.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.25.11
uvicorn==0.17.6
yarl==1.8.2
